List editing
Cleanup of the forms list editing UI
Moves icons back to the right
Removes the edit list followed by toggle to edit item flow. Instead you can just hover on a list and click the edit (or delete or new item)
Moves the edit icon for the text edit to the right, using the widget itself to render it so it can appear right next to the text.
Adds header to the list view, which you can control via a static member on the item's form.