Refactored controls that are rendered from the controls declared by the user
For example, the user would add a TextEdit control to a form, but when that is rendered it has a set of props added to it. At that point it becomes a TextEditControl (wrapped in a formGroup). This lets us be clearer about types.

Also added a ChooserControl etc. Still some work to do on that, it's just the most basic implementation with the new react-select api.