Async choosers
Adds displayView prop which can either be a string or a function that returns a React Element. Like the previous view prop, this is used to show specialized rendering for the chooser when data is not loaded yet.

Added an example of the chooser async loading data when the inline editing is triggered and then selecting by search from a large list.